backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector database client
qdrant-client==1.7.0

# LLM and embedding dependencies
llama-cpp-python==0.2.19
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# OpenAI API support
openai==1.3.7

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3

# HTTP client for API calls
httpx==0.25.2
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Security
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2

# File handling
python-magic==0.4.27